Don't reduce scalars when converting
The operator "scalar * product" will pre-multiply the scalar with the
first term in 'product' if that term is also a scalar (because, for
example, we want "5 * 4L" to result in "20L", not "5 4 L").  However,
we don't want this behavior for unit conversions.  For example, if
we are asked to convert "2L" to "1|4 cups", we want the result in
multiples of quarter cups, not whole cups.