Accept names that end in multiple digits without _
Names ending in multiple digits without a preceding underscore seems to
be forbidden by the GNU Units spec:

http://www.gnu.org/software/units/manual/units.html#Defining-New-Units

  "Names cannot begin with a digit, and if a name ends in a digit other
  than zero, the digit must be preceded by a string beginning with an
  underscore, and afterwords consisting of only digits, decimal points
  or commas."

Nevertheless, the units definition file that comes preinstalled with GNU
Units (version 2.00) has the following definition:

    bronzeyard11            914.39980 mm

So we'll just accept unit names ending in multiple digits.  This should
not conflict with the exponentiation operator (e.g., "m2" for square
meters) because the GNU Units spec also says that only single digits may
be used in this way.